Refactor themes to src
Move ThemeWrapper to App from server and client
