+++ src/components/PaginationComponent/__tests__/PaginationComponent.test.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { createMount } from '@material-ui/core/test-utils';
 import { MuiThemeProvider } from '@material-ui/core';
 import { IntlProvider } from 'react-intl';
-import themes from '../../../../themes';
+import themes from '../../../themes';
 import PaginationComponent from '../index';
 
 // Mock props for intl provider
