+++ src/components/ServiceMapButton/__tests__/ServiceMapButton.test.js
@@ -4,7 +4,7 @@ import { IntlProvider } from 'react-intl';
 import { createMount } from '@material-ui/core/test-utils';
 import { MuiThemeProvider } from '@material-ui/core';
 import ServiceMapButton from '../index';
-import themes from '../../../../themes';
+import themes from '../../../themes';
 
 // Mock props for intl provider
 const intlMock = {
